Correct libusb usage on FreeBSD (#10736)
Change type of Handle on FreeBSD. On FreeBSD the libusb_hotplug_register_callback() function uses a pointer to a struct as a handle.

---------

+++ src/gui/osutils/DeviceListener.cpp
@@ -57,7 +57,7 @@ DeviceListener::registerHotplugCallback(bool arrived, bool left, int vendorId, i
 void DeviceListener::deregisterHotplugCallback(Handle handle)
 {
 #if defined(Q_OS_UNIX) && !defined(Q_OS_MACOS)
-    m_listeners[0]->deregisterHotplugCallback(static_cast<int>(handle));
+    m_listeners[0]->deregisterHotplugCallback(handle);
 #else
     if (m_listeners.contains(handle)) {
         m_listeners[handle]->deregisterHotplugCallback();
